Fixed crash when saving new character. Old POSIX rename() would work fine even if the file doesn't exist. But C++17 fs:rename() throws exception.

+++ src/repos/playerrepository.cpp
@@ -845,7 +845,11 @@ void InMemoryPlayerRepository::Save(const Character *pc) const
 
     pc->PCData->SaveTime = current_time;
 
-    fs::rename(GetPlayerFilename(pc).c_str(), GetPlayerBackupFilename(pc->Name).c_str());
+    if(fs::exists(GetPlayerFilename(pc)))
+    {
+        fs::rename(GetPlayerFilename(pc).c_str(), GetPlayerBackupFilename(pc->Name).c_str());
+    }
+    
     LuaSaveDataFile(GetPlayerFilename(pc),
         &InMemoryPlayerRepository::PushPlayer,
         "character", pc);
